+++ apps/www/widgets/exchanges/card/delete.button.tsx
@@ -26,10 +26,10 @@ export function Exchange_Delete_Button() {
   return (
     <Button
       variant={{ intent: "light", size: "sm", state: "ghost" }}
-      className="box-content h-4 py-2"
+      className="box-content h-7 p-2"
       onPress={() => set_outlet({ state: "should_delete" })}
     >
-      <Trash className="h-full text-white" />
+      <Trash className="h-4 text-white" />
     </Button>
   );
 }
